adds match score based on relevance

+++ testing_webpage/dashboard/models.py
@@ -48,7 +48,7 @@ class Candidate(models.Model):
     comments = models.ManyToManyField(Comment, default=[])
     evaluations = models.ManyToManyField(Evaluation)
     surveys = models.ManyToManyField(Survey)
-    match = models.IntegerField(null=True, default=None)
+    match = models.FloatField(null=True, default=None)
 
     created_at = models.DateTimeField(auto_now_add=True)
     updated_at = models.DateTimeField(auto_now=True)
